// src/app/models/admin/dangkyhoatdongkhoangsan/dkthamdokhuvuc.model.ts
export class OutputDkThamDoKhuVucModel {
    idthamdokhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    iddangkythamdo: string;
    loaikhuvuc: number;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    siteid: string;
}

export class InputDkThamDoKhuVucModel {
    idthamdokhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    iddangkythamdo: string;
    loaikhuvuc: number;
    hequychieu: string;
}